Chatbot/requirements.txt
transformers
adb-shell
Pillow
